requirements.txt
asttokens==3.0.0
backcall==0.2.0
bleach==6.1.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
cmake==3.25.0
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
entrypoints==0.4
executing==2.2.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
grad-cam==1.5.4
h5py==3.11.0
huggingface-hub==0.25.1
idna==3.10
importlib_resources==6.4.5
ipykernel==6.29.5
# ipython==9.2.0
# ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.6.17
kiwisolver==1.4.7
lit==15.0.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==25.0
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.7
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytorch-gradcam==0.2.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.3.2
scikit-multilearn==0.2.0
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
text-unidecode==1.3
threadpoolctl==3.5.0
timm==1.0.9
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
triton==2.0.0
ttach==0.0.3
typing_extensions==4.13.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
zipp==3.20.2
